# builtin: speed up string concatenation and `repeat()` method with vme mcpy