@i582 2023-05-19
> Maybe your PR covers #276 ?

It seems not, it suggests optimizing several concatenations in a row in order to allocate memory only once